examples/a2aclient/requirements.txt
a2a-sdk==0.3.0
httpx>=0.24,<1.0